+++ extensions/omni/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2024-11-30
+
+### Changed
+
+* Backend initialization transaction ends with a commit [#699](https://github.com/omnigres/omnigres/pull/699)
+
 ## [0.2.2] - 2024-11-29
 
 ### Changed
@@ -96,3 +102,5 @@ Initial release following a few months of iterative development.
 [0.2.1]: [https://github.com/omnigres/omnigres/pull/677]
 
 [0.2.2]: [https://github.com/omnigres/omnigres/pull/698]
+
+[0.2.3]: [https://github.com/omnigres/omnigres/pull/699]