+++ extensions/omni_httpd/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2024-11-30
+
+### Added
+
+* Explicit omni_httpd server controls [#699](https://github.com/omnigres/omnigres/pull/699)
+
 ## [0.2.3] - 2024-11-29
 
 ### Fixed
@@ -76,3 +82,5 @@ Initial release following a few months of iterative development.
 [0.2.2]: [https://github.com/omnigres/omnigres/pull/692]
 
 [0.2.3]: [https://github.com/omnigres/omnigres/pull/698]
+
+[0.2.4]: [https://github.com/omnigres/omnigres/pull/699]